Extract out the list view toggling code in main window
